test(wallet): add common test module for macros
